Start integrating crash dump info into MLN
